visable borders
Newb here, I made a simple border function. This is my first pull request. I spent some time trying to make the lines end at the corners but decided to keep it simple. I think it would be nice to be able to toggle these borders. Thanks for creating this project.